Stop using deprecated Django APIs
